+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/mvc/tobe/adapter/HandlerAdapters.java
@@ -1,0 +1,24 @@
+package webmvc.org.springframework.web.servlet.mvc.tobe.adapter;
+
+import jakarta.servlet.ServletException;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class HandlerAdapters {
+
+    private final List<HandlerAdapter> handlerAdapters = new ArrayList<>();
+
+    public void addHandlerAdapter(HandlerAdapter handlerAdapter) {
+        handlerAdapters.add(handlerAdapter);
+    }
+
+    public HandlerAdapter getHandlerAdapter(Object handler) throws ServletException {
+        for (HandlerAdapter handlerAdapter : handlerAdapters) {
+            if (handlerAdapter.supports(handler)) {
+                return handlerAdapter;
+            }
+        }
+        throw new ServletException("handlerAdapter not found for handler: " + handler);
+    }
+}
